backend/requirements.txt
fastapi~=0.116.1
uvicorn~=0.35.0
colorama~=0.4.6
python-dotenv~=1.1.1
requests~=2.32.4
PyJWT~=2.10.1